+++ changelog.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 1.0.0 
+
+- Preapration to foundryvtt 0.8.0
+- Converted to typescript project
+- Direct integration with [Token Auras](https://bitbucket.org/Fyorl/token-auras/src/master/) ty to  [Fyorl](https://bitbucket.org/Fyorl/)
+- Direct integration with [NoTokenAnim](https://bitbucket.org/Fyorl/notokenanim/src/master/) ty to  [Fyorl](https://bitbucket.org/Fyorl/)
+- Direct integration with [Colored Effects](https://github.com/sean8223/coloredeffects) ty to [sean8223](https://github.com/sean8223)
+
 ## 0.0.8
 
 - No functional changes, converted the javascript project to typescript, some minor bug fix with FoundryVTT release 0.7.9.